+++ src/main/java/io/github/forgestove/create_cyber_goggles/CCG.java
@@ -35,7 +35,7 @@ public class CCG {
 		@SubscribeEvent
 		public static void clientTickEvent(ClientTickEvent event) {
 			KineticDebugger.tick();
-			AirBoxRender.tick();
+			DelayRender.tick();
 		}
 	}
 	@EventBusSubscriber(modid = ID, value = Dist.CLIENT, bus = Bus.MOD)