+++ src/main/java/io/github/forgestove/create_cyber_goggles/CCGConfig.java
@@ -1,19 +1,18 @@
 package io.github.forgestove.create_cyber_goggles;
-import io.github.forgestove.create_cyber_goggles.CCGConfig.Goggles.GameMode;
 import me.shedaniel.autoconfig.*;
 import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.autoconfig.annotation.ConfigEntry.Category;
+import me.shedaniel.autoconfig.annotation.ConfigEntry.*;
 import me.shedaniel.autoconfig.annotation.ConfigEntry.Gui.*;
 import net.minecraftforge.client.ConfigScreenHandler.ConfigScreenFactory;
 import net.minecraftforge.fml.ModLoadingContext;
 @Config(name = CCG.ID)
 public class CCGConfig implements ConfigData {
 	@Category("goggles") @TransitiveObject public final Goggles goggles = new Goggles();
-	@Category("goggles") @CollapsibleObject public final GameMode gameMode = new GameMode();
+	@Category("gameMode") @TransitiveObject public final GameMode gameMode = new GameMode();
+	@Category("delayRender") @TransitiveObject public final DelayRender delayRender = new DelayRender();
 	@Category("chainConveyor") @TransitiveObject public final ChainConveyor chainConveyor = new ChainConveyor();
-	@Category("armor") @TransitiveObject public final Armor armor = new Armor();
 	@Category("wrench") @TransitiveObject public final Wrench wrench = new Wrench();
-	@Category("other") @TransitiveObject public final Other other = new Other();
+	@Category("misc") @TransitiveObject public final Misc misc = new Misc();
 	public static void register() {
 		var factory = new ConfigScreenFactory((mc, screen) -> AutoConfig.getConfigScreen(CCGConfig.class, screen).get());
 		ModLoadingContext.get().registerExtensionPoint(ConfigScreenFactory.class, () -> factory);
@@ -24,16 +23,21 @@ public class CCGConfig implements ConfigData {
 		@Tooltip public boolean betterStoreInfo = true;
 		@Tooltip public boolean renderExtraItems = true;
 		@Tooltip public boolean enableKineticEffect = true;
-		@Tooltip public boolean preciseNumbers = true;
+		@Tooltip public boolean preciseNumber = true;
 		@Tooltip public boolean disableScreenGoggles = true;
 		@Tooltip public boolean betterLine = true;
-		@Tooltip public boolean renderBox = true;
-		public static class GameMode {
-			@Tooltip public boolean enableInSurvival = true;
-			@Tooltip public boolean enableInCreative = true;
-			@Tooltip public boolean enableInSpectator = true;
-			@Tooltip public boolean enableInAdventure = true;
-		}
+	}
+	public static class GameMode {
+		@Tooltip public boolean enableInSurvival = true;
+		@Tooltip public boolean enableInCreative = true;
+		@Tooltip public boolean enableInSpectator = true;
+		@Tooltip public boolean enableInAdventure = true;
+	}
+	public static class DelayRender {
+		@Tooltip public boolean renderAnalogBox = true;
+		@Tooltip public int delayRenderDuration = 60;
+		@Tooltip @ColorPicker public int airBoxPushColor = 0xDDC166;
+		@Tooltip @ColorPicker public int airBoxPullColor = 0x7FCDE0;
 	}
 	public static class ChainConveyor {
 		@Tooltip public boolean alwaysAllowRiding = false;
@@ -41,21 +45,19 @@ public class CCGConfig implements ConfigData {
 		@Tooltip public boolean enhancedConnection = true;
 		@Tooltip public boolean cardBoardedYourself = false;
 	}
-	public static class Armor {
-		@Tooltip public boolean removeBoxOverlay = true;
-		@Tooltip public boolean removeNetheriteFirstPerson = false;
-		@Tooltip public boolean removeDivingBootsAffect = false;
-	}
 	public static class Wrench {
 		@Tooltip public boolean alwaysAllowRotating = true;
 		@Tooltip public boolean removeCooldown = true;
 	}
-	public static class Other {
+	public static class Misc {
+		@Tooltip public boolean removeCardboardOverlay = true;
+		@Tooltip public boolean removeNetheriteFirstPerson = false;
+		@Tooltip public boolean removeDivingBootsAffect = false;
 		@Tooltip public boolean fixSchematicName = true;
 		@Tooltip public boolean rightClickPenetrate = false;
 		@Tooltip public boolean rainbowDebug = false;
 		@Tooltip public boolean forcedBackend = false;
-		@Tooltip @RequiresRestart public boolean showScrapContent = true;
 		@Tooltip public boolean nbtFix = false;
+		@Tooltip @RequiresRestart public boolean showScrapContent = true;
 	}
 }