+++ src/main/java/io/github/forgestove/create_cyber_goggles/mixin/misc/FilesHelperMixin.java
@@ -1,4 +1,4 @@
-package io.github.forgestove.create_cyber_goggles.mixin.other;
+package io.github.forgestove.create_cyber_goggles.mixin.misc;
 import com.simibubi.create.foundation.utility.FilesHelper;
 import io.github.forgestove.create_cyber_goggles.CCG;
 import net.createmod.catnip.lang.Lang;
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class FilesHelperMixin {
 	@Inject(method = "slug", at = @At("HEAD"), cancellable = true)
 	private static void slug(String name, CallbackInfoReturnable<String> returnable) {
-		if (!CCG.CONFIG.other.fixSchematicName) return;
+		if (!CCG.CONFIG.misc.fixSchematicName) return;
 		returnable.setReturnValue(Lang.asId(name).replaceAll("[\\\\/:*?\"<>|]+", "_"));
 	}
 }