+++ src/main/java/io/github/forgestove/create_cyber_goggles/mixin/misc/SimpleBackendMixin.java
@@ -1,4 +1,4 @@
-package io.github.forgestove.create_cyber_goggles.mixin.other;
+package io.github.forgestove.create_cyber_goggles.mixin.misc;
 import dev.engine_room.flywheel.lib.backend.SimpleBackend;
 import io.github.forgestove.create_cyber_goggles.CCG;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,6 +8,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class SimpleBackendMixin {
 	@Inject(method = "isSupported", at = @At("HEAD"), cancellable = true)
 	public void isSupported(CallbackInfoReturnable<Boolean> returnable) {
-		if (CCG.CONFIG.other.forcedBackend) returnable.setReturnValue(true);
+		if (CCG.CONFIG.misc.forcedBackend) returnable.setReturnValue(true);
 	}
 }